+++ lib/src/clojure_lsp/feature/refactor.clj
@@ -18,8 +18,8 @@
 (defmethod refactor :add-import-to-namespace [{:keys [loc args db]}]
   (apply f.add-missing-libspec/add-import-to-namespace loc (concat args [db])))
 
-(defmethod refactor :add-missing-libspec [{:keys [loc db]}]
-  (f.add-missing-libspec/add-missing-libspec loc db))
+(defmethod refactor :add-missing-libspec [{:keys [loc uri db]}]
+  (f.add-missing-libspec/add-missing-libspec loc uri db))
 
 (defmethod refactor :add-require-suggestion [{:keys [loc args db]}]
   (apply f.add-missing-libspec/add-require-suggestion loc (concat args [db])))