+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - parser: more efficiently seek to cursor position, improving performance especially in large files. #793 @mainej
   - Fix clean-ns not sorting properly node requires for cljs. #815
   - Fix move-to-let to ensure locals don't move out of scope. #830
+  - Improve logic around require suggestions. #837
 
 - Editor
   - extract-function: Fix wrong args when extracting from multi-arity fn. #683